fix crash...?
after running for a while, app crashes
having crashed once, seems crash is more likely to recur
hooked up to debugger, device disconnects upon crash

restarting ios and host seems to have fixed the disconnect
it may have also fixed the crash, and the rest is something that never happened

next, if the 5s delay wasn't enough for takePicture to work,
upon next present of picker (which is still presented), crash

a couple suggestions from internet didn't work:
1) subscribe to AVCaptureSessionDidStartRunning
2) completion block of presenting picker

-----------------------------------------------------
an interlude

AVCaptureSessionDidStartRunning is referenced like this:
`NSNotification.Name.AVCaptureSessionDidStartRunning`
and you have to do
`import AVFoundation`
to get access to it.

Wait, what?

https://developer.apple.com/documentation/foundation/nsnotification/name
`static let AVCaptureDeviceWasConnected: NSNotification.Name`
what? so it's in there, right? what?

https://developer.apple.com/documentation/foundation/nsnotification/name/1387835-avcapturesessiondidstartrunning#declarations
am I supposed to know to import based on the URL?

thanks for thinking different, apple
-----------------------------------------------------

so stuff it, let's just take a picture every 10 seconds
it'll either get ignored because the picker isn't presented
or ignored because the camera isn't ready
but once it's not ignored, everything goes back to normal,
and this should prevent any crashes

time will tell.